unfinished changes

+++ src/localization/filters/terrain_particle_filter.hpp
@@ -31,8 +31,8 @@ private:
 
 public:
     TerrainParticleFilter(const std::string& terrainFilename, double sigmaX, double sigmaTheta, const Eigen::Vector2d& footprint);   
-    void init_particles(const manif::SE2d& initialPose, int numParticles);
-    void init_particles(int numParticles);
+    void init_particles(const manif::SE2d& initialPose, size_t numParticles);
+    void init_particles(size_t numParticles);
 
     // TODO: add overloads for odometry and IMU pose
     void predict(const Eigen::Vector3d& velCmd, double dt);
